requirements.txt
torch
torchvision
torchaudio
opencv-python
numpy
mediapipe
hydra-core
pytorch-lightning
sentencepiece
omegaconf
